requirements.txt
# AWS library
boto3

# Dataframe libaray
pandas

# publish subscribe data streaming
pubnub

# .env handler
python-dotenv

# cluster library
pyspark

# tweet classification
textblob

googletrans

# kafka data streaming library
kafka-python

requests

# machine learning library
scikit-learn